+++ src/constants/header.ts
@@ -1,0 +1,3 @@
+export const HEADER_HEIGHT = 80;
+export const SCROLL_THRESHOLD = 0.3;
+export const MOBILE_MENU_WIDTH = 200;
